메시지 결합 완료

+++ src/main/java/com/petstagram/dto/ChatRoomDTO.java
@@ -23,7 +23,7 @@ public class ChatRoomDTO {
     private String senderName; // 발신자 이름
     private Long receiverId; // 수신자 ID
     private String receiverName; // 수신자 이름
-    private int unreadMessageCount;
+    private Long unreadMessageCount;
 
     // Entity -> DTO 변환 메서드
     public static ChatRoomDTO toDTO(ChatRoomEntity chatRoom) {
