+++ src/main/java/com/petstagram/repository/MessageRepository.java
@@ -17,9 +17,9 @@ public interface MessageRepository extends JpaRepository<MessageEntity, Long> {
 
     // 채팅방 ID와 사용자 ID에 따른 읽지 않은 메시지 개수를 계산하는 쿼리 메서드
     @Query("SELECT COUNT(m) FROM MessageEntity m WHERE m.chatRoom.id = :chatRoomId AND m.isRead = false AND m.sender.id != :currentUserId")
-    int countUnreadMessages(@Param("chatRoomId") Long chatRoomId, @Param("currentUserId") Long currentUserId);
+    Long countUnreadMessages(@Param("chatRoomId") Long chatRoomId, @Param("currentUserId") Long currentUserId);
 
     // 사용자의 모든 읽지 않은 메시지 개수를 계산하는 쿼리 메서드
     @Query("SELECT COUNT(m) FROM MessageEntity m WHERE m.receiver.id = :userId AND m.isRead = false")
-    int countUnreadMessagesForUser(@Param("userId") Long userId);
+    Long countUnreadMessagesForUser(@Param("userId") Long userId);
 }
